+++ cc-core/controllers/account/update_profile.php
@@ -15,63 +15,77 @@ $tempFilePrefix = UPLOAD_PATH . '/temp/' . $this->view->vars->loggedInUser->user
 // Update profile if requested
 if (isset($_POST['submitted'])) {
 
-    $this->view->vars->profile_submit = true;
+    // Validate form nonce token and submission speed
+    if (
+        !empty($_POST['nonce'])
+        && !empty($_SESSION['formNonce'])
+        && !empty($_SESSION['formTime'])
+        && $_POST['nonce'] == $_SESSION['formNonce']
+        && time() - $_SESSION['formTime'] >= 3
+    ) {
 
-    // Validate First Name
-    if (!empty($this->view->vars->loggedInUser->firstName) && $_POST['first_name'] == '') {
-        $this->view->vars->loggedInUser->firstName = '';
-    } elseif (!empty($_POST['first_name'])) {
-        $this->view->vars->loggedInUser->firstName = trim($_POST['first_name']);
-    }
+        $this->view->vars->profile_submit = true;
 
-    // Validate Last Name
-    if (!empty($this->view->vars->loggedInUser->lastName) && $_POST['last_name'] == '') {
-        $this->view->vars->loggedInUser->lastName = '';
-    } elseif (!empty($_POST['last_name'])) {
-        $this->view->vars->loggedInUser->lastName = trim($_POST['last_name']);
-    }
-
-    // Validate Email
-    if (!empty($_POST['email']) && preg_match('/^[a-z0-9][a-z0-9_\.\-]+@[a-z0-9][a-z0-9\.\-]+\.[a-z0-9]{2,4}$/i', $_POST['email'])) {
-        $userCheck = $userMapper->getUserByCustom(array('email' => $_POST['email']));
-        if (!$userCheck || $userCheck->email == $this->view->vars->loggedInUser->email) {
-            $this->view->vars->loggedInUser->email = $_POST['email'];
-        } else {
-            $this->view->vars->Errors['email'] = Language::GetText('error_email_unavailable');
+        // Validate First Name
+        if (!empty($this->view->vars->loggedInUser->firstName) && $_POST['first_name'] == '') {
+            $this->view->vars->loggedInUser->firstName = '';
+        } elseif (!empty($_POST['first_name'])) {
+            $this->view->vars->loggedInUser->firstName = trim($_POST['first_name']);
         }
-    } else {
-        $this->view->vars->Errors['email'] = Language::GetText('error_email');
-    }
 
-    // Validate Website
-    if (!empty($_POST['website'])) {
-        $website = $_POST['website'];
-        if (preg_match ('/^(https?:\/\/)?[a-z0-9][a-z0-9\.-]+\.[a-z0-9]{2,4}.*$/i', $website, $matches)) {
-            $website = (empty($matches[1]) ? 'http://' : '') . $website;
-            $this->view->vars->loggedInUser->website = trim($website);
-        } else {
-            $this->view->vars->errors['website'] = Language::GetText('error_website_invalid');
+        // Validate Last Name
+        if (!empty($this->view->vars->loggedInUser->lastName) && $_POST['last_name'] == '') {
+            $this->view->vars->loggedInUser->lastName = '';
+        } elseif (!empty($_POST['last_name'])) {
+            $this->view->vars->loggedInUser->lastName = trim($_POST['last_name']);
         }
-    } else {
-        $this->view->vars->loggedInUser->website = '';
-    }
 
-    // Validate About Me
-    if (!empty($this->view->vars->loggedInUser->aboutMe) && $_POST['about_me'] == '') {
-        $this->view->vars->loggedInUser->aboutMe = '';
-    } elseif (!empty($_POST['about_me'])) {
-        $this->view->vars->loggedInUser->aboutMe = trim($_POST['about_me']);
-    }
+        // Validate Email
+        if (!empty($_POST['email']) && preg_match('/^[a-z0-9][a-z0-9_\.\-]+@[a-z0-9][a-z0-9\.\-]+\.[a-z0-9]{2,4}$/i', $_POST['email'])) {
+            $userCheck = $userMapper->getUserByCustom(array('email' => $_POST['email']));
+            if (!$userCheck || $userCheck->email == $this->view->vars->loggedInUser->email) {
+                $this->view->vars->loggedInUser->email = $_POST['email'];
+            } else {
+                $this->view->vars->Errors['email'] = Language::GetText('error_email_unavailable');
+            }
+        } else {
+            $this->view->vars->Errors['email'] = Language::GetText('error_email');
+        }
 
-    // Update User if no errors were found
-    if (empty ($this->view->vars->Errors)) {
-        $this->view->vars->message = Language::GetText('success_profile_updated');
-        $this->view->vars->message_type = 'success';
-        $userMapper->save($this->view->vars->loggedInUser);
+        // Validate Website
+        if (!empty($_POST['website'])) {
+            $website = $_POST['website'];
+            if (preg_match ('/^(https?:\/\/)?[a-z0-9][a-z0-9\.-]+\.[a-z0-9]{2,4}.*$/i', $website, $matches)) {
+                $website = (empty($matches[1]) ? 'http://' : '') . $website;
+                $this->view->vars->loggedInUser->website = trim($website);
+            } else {
+                $this->view->vars->errors['website'] = Language::GetText('error_website_invalid');
+            }
+        } else {
+            $this->view->vars->loggedInUser->website = '';
+        }
+
+        // Validate About Me
+        if (!empty($this->view->vars->loggedInUser->aboutMe) && $_POST['about_me'] == '') {
+            $this->view->vars->loggedInUser->aboutMe = '';
+        } elseif (!empty($_POST['about_me'])) {
+            $this->view->vars->loggedInUser->aboutMe = trim($_POST['about_me']);
+        }
+
+        // Update User if no errors were found
+        if (empty ($this->view->vars->Errors)) {
+            $this->view->vars->message = Language::GetText('success_profile_updated');
+            $this->view->vars->message_type = 'success';
+            $userMapper->save($this->view->vars->loggedInUser);
+        } else {
+            $this->view->vars->message = Language::GetText('errors_below');
+            $this->view->vars->message .= '<br /><br /> - ' . implode ('<br /> - ', $this->view->vars->Errors);
+            $this->view->vars->message_type = 'errors';
+        }
+
     } else {
-        $this->view->vars->message = Language::GetText('errors_below');
-        $this->view->vars->message .= '<br /><br /> - ' . implode ('<br /> - ', $this->view->vars->Errors);
-        $this->view->vars->message_type = 'errors';
+        $this->view->vars->message = Language::getText('invalid_session');
+        $this->view->vars->message_type = 'alert-danger';
     }
 }
 
@@ -99,29 +113,48 @@ if (
     && \App::isValidUpload($_POST['upload']['temp'], $this->view->vars->loggedInUser, 'image')
 ) {
 
-    $this->view->vars->avatar_submit = true;
-    $filename = Avatar::generateFilename();
-    $avatar = UPLOAD_PATH . '/avatars/' . $filename . '.png';
+    // Validate form nonce token and submission speed
+    if (
+        !empty($_POST['nonce'])
+        && !empty($_SESSION['formNonce'])
+        && !empty($_SESSION['formTime'])
+        && $_POST['nonce'] == $_SESSION['formNonce']
+        && time() - $_SESSION['formTime'] >= 3
+    ) {
 
-    // Attempt to save avatar
-    if (Avatar::saveAvatar($_POST['upload']['temp'], $filename)) {
+        $this->view->vars->avatar_submit = true;
+        $filename = Avatar::generateFilename();
+        $avatar = UPLOAD_PATH . '/avatars/' . $filename . '.png';
 
-        // Check for existing avatar
-        if (!empty($this->view->vars->loggedInUser->avatar)) Avatar::delete($this->view->vars->loggedInUser->avatar);
+        // Attempt to save avatar
+        if (Avatar::saveAvatar($_POST['upload']['temp'], $filename)) {
 
-        // Update User
-        $userMapper = new UserMapper();
-        $this->view->vars->loggedInUser->avatar = $filename;
-        $userMapper->save($this->view->vars->loggedInUser);
+            // Check for existing avatar
+            if (!empty($this->view->vars->loggedInUser->avatar)) Avatar::delete($this->view->vars->loggedInUser->avatar);
 
-        $this->view->vars->message = Language::getText('success_profile_updated');
-        $this->view->vars->message_type = 'success';
-        $userMapper->save($this->view->vars->loggedInUser);
+            // Update User
+            $userMapper = new UserMapper();
+            $this->view->vars->loggedInUser->avatar = $filename;
+            $userMapper->save($this->view->vars->loggedInUser);
+
+            $this->view->vars->message = Language::getText('success_profile_updated');
+            $this->view->vars->message_type = 'success';
+            $userMapper->save($this->view->vars->loggedInUser);
+
+        } else {
+            $this->view->vars->message = Language::getText('error_upload_system', array ('host' => HOST));
+            $this->view->vars->message_type = 'errors';
+        }
 
     } else {
-        $this->view->vars->message = Language::getText('error_upload_system', array ('host' => HOST));
-        $this->view->vars->message_type = 'errors';
+        $this->view->vars->message = Language::getText('invalid_session');
+        $this->view->vars->message_type = 'alert-danger';
     }
 }
+
+// Generate new form nonce
+$this->view->vars->formNonce = md5(uniqid(rand(), true));
+$_SESSION['formNonce'] = $this->view->vars->formNonce;
+$_SESSION['formTime'] = time();
 
 Plugin::triggerEvent('update_profile.end');
